+++ feature/notification/src/main/java/com/expenseanalyst/feature/notification/parser/BillStatementParserRegistry.kt
@@ -14,6 +14,7 @@ object BillStatementParserRegistry {
         EmiratesNbdStatementParser(),
         AlRajhiStatementParser(),
         HdfcStatementParser(),
+        OneCardStatementParser(),
         TamaraStatementParser(),
         SaudiEnergyStatementParser(),
         EjarStatementParser(),